+++ NEWS.md
@@ -7,6 +7,8 @@
       intervals are placed next to parameter estimates.
     - Add `vcov` method for `std_selected` output. If bootstrapping is used,
       it can return the variance-covariance matrix of the bootstrap estimates.
+    - Add `confint` method for `std_selected` output. If bootstrapping is used,
+      it can return the bootstrap percentile confidence intervals if requested. 
 
 - 0.1.3